backend/requirements.txt
arabic-reshaper
beautifulsoup4
blinker
certifi
cffi
chardet
charset-normalizer
click
colorama
cryptography
deep-translator
Deprecated
filelock
Flask
fonttools
fpdf
fsspec
googletrans
h11
h2
hpack
hstspreload
httpcore
httpx
huggingface-hub
hyperframe
idna
itsdangerous
Jinja2
lxml
MarkupSafe
mpmath
networkx
numpy
packaging
pdf2image
pdfminer.six
pdfplumber
pdfrw
pikepdf
pillow
pycparser
PyMuPDF
pypdf
PyPDF2
pypdfium2
pytesseract
python-bidi
PyYAML
regex
reportlab
requests
rfc3986
safetensors
sniffio
soupsieve
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
Werkzeug
wrapt
python-multipart
fastapi
fastapi[standard]
uvicorn[standard]
argostranslate
wordninja
spacy